fix behavior and lint
